requirements.txt
slackclient==2.5.0
python-dotenv==0.11.0
